+++ crates/orchestrator/src/jobs/snos_job/mod.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
 use async_trait::async_trait;
+use chrono::{SubsecRound, Utc};
 use color_eyre::Result;
 use uuid::Uuid;
 
@@ -28,6 +29,8 @@ impl Job for SnosJob {
             external_id: String::new().into(),
             metadata,
             version: 0,
+            created_at: Utc::now().round_subsecs(0),
+            updated_at: Utc::now().round_subsecs(0),
         })
     }