+++ crates/orchestrator/src/jobs/types.rs
@@ -4,8 +4,9 @@ use color_eyre::eyre::eyre;
 use color_eyre::Result;
 use da_client_interface::DaVerificationStatus;
 // TODO: job types shouldn't depend on mongodb
+use chrono::{DateTime, Utc};
 #[cfg(feature = "with_mongodb")]
-use mongodb::bson::serde_helpers::uuid_1_as_binary;
+use mongodb::bson::serde_helpers::{chrono_datetime_as_bson_datetime, uuid_1_as_binary};
 use serde::{Deserialize, Serialize};
 use settlement_client_interface::SettlementVerificationStatus;
 use uuid::Uuid;
@@ -133,6 +134,12 @@ pub struct JobItem {
     pub metadata: HashMap<String, String>,
     /// helps to keep track of the version of the item for optimistic locking
     pub version: i32,
+    /// timestamp when the job was created
+    #[cfg_attr(feature = "with_mongodb", serde(with = "chrono_datetime_as_bson_datetime"))]
+    pub created_at: DateTime<Utc>,
+    /// timestamp when the job was last updated
+    #[cfg_attr(feature = "with_mongodb", serde(with = "chrono_datetime_as_bson_datetime"))]
+    pub updated_at: DateTime<Utc>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]