+++ crates/orchestrator/src/tests/database/mod.rs
@@ -2,6 +2,7 @@ use crate::config::{config, Config};
 use crate::jobs::types::{ExternalId, JobItem, JobStatus, JobType};
 use crate::tests::config::TestConfigBuilder;
 use arc_swap::Guard;
+use chrono::{SubsecRound, Utc};
 use rstest::*;
 use std::sync::Arc;
 use uuid::Uuid;
@@ -224,5 +225,7 @@ pub fn build_job_item(job_type: JobType, job_status: JobStatus, internal_id: u64
         external_id: ExternalId::Number(0),
         metadata: Default::default(),
         version: 0,
+        created_at: Utc::now().round_subsecs(0),
+        updated_at: Utc::now().round_subsecs(0),
     }
 }