+++ crates/orchestrator/src/tests/jobs/da_job/mod.rs
@@ -6,6 +6,7 @@ use crate::tests::common::drop_database;
 use crate::tests::config::TestConfigBuilder;
 use crate::{config::config, jobs::Job};
 use assert_matches::assert_matches;
+use chrono::{SubsecRound, Utc};
 use color_eyre::eyre::eyre;
 use da_client_interface::MockDaClient;
 use mockall::predicate::always;
@@ -68,6 +69,8 @@ async fn test_da_job_process_job_failure_on_small_blob_size(
                 external_id: ExternalId::String(internal_id.to_string().into_boxed_str()),
                 metadata: HashMap::default(),
                 version: 0,
+                created_at: Utc::now().round_subsecs(0),
+                updated_at: Utc::now().round_subsecs(0),
             },
         )
         .await;
@@ -126,6 +129,8 @@ async fn test_da_job_process_job_failure_on_pending_block() {
                 external_id: ExternalId::String("1".to_string().into_boxed_str()),
                 metadata: HashMap::default(),
                 version: 0,
+                created_at: Utc::now().round_subsecs(0),
+                updated_at: Utc::now().round_subsecs(0),
             },
         )
         .await;
@@ -202,6 +207,8 @@ async fn test_da_job_process_job_success(
                 external_id: ExternalId::String(internal_id.to_string().into_boxed_str()),
                 metadata: HashMap::default(),
                 version: 0,
+                created_at: Utc::now().round_subsecs(0),
+                updated_at: Utc::now().round_subsecs(0),
             },
         )
         .await;