+++ crates/orchestrator/src/tests/jobs/mod.rs
@@ -17,6 +17,7 @@ pub mod proving_job;
 pub mod state_update_job;
 
 use assert_matches::assert_matches;
+use chrono::{SubsecRound, Utc};
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::Duration;
@@ -505,6 +506,8 @@ fn build_job_item_by_type_and_status(job_type: JobType, job_status: JobStatus, i
         external_id: ExternalId::Number(0),
         metadata: hashmap,
         version: 0,
+        created_at: Utc::now().round_subsecs(0),
+        updated_at: Utc::now().round_subsecs(0),
     }
 }